Simplify JSON --provider handling
When the user passes "--provider <JSON-VALUE>", rather than reading it
into an e2e.CloudConfig, autodetecting the
exutilcluster.ClusterConfig, and then merging the two structs together
field by field, just do the autodetection first, and then unmarshal
the JSON directly into the ClusterConfig (possibly overwriting some of
the autodetected values).

In addition to being simpler, this allows for overriding the
networking-related values that are in ClusterConfig but not
CloudConfig.